Add Wido widget (#34)
A new button it's being added on the header action bar. It opens a
dialog that contains the widget to bridge assets from Ethereum.

`web3-react` dependencies have been added in order to support for most
of wallet providers on the Ethereum ecosystem.
`etheres` has been downgraded to v5. v6 is relatively new and the widget
doesn't support it yet. I checked and there were no hard dependencies on
the v6.